Karolis_Živulskas/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repops"
version = "0.1.0"
description = "Reputation Operations — social media disinformation & hate speech monitoring platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Collection
    "playwright>=1.44",
    "httpx>=0.27",
    "tenacity>=8.3",

    # NLP / Analysis
    "pyahocorasick>=2.0",

    # Task queue
    "celery[redis]>=5.4",

    # Storage
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.1",
    "elasticsearch>=8.13",
    "boto3>=1.34",

    # API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-multipart>=0.0.9",

    # Observability
    "structlog>=24.0",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "factory-boy>=3.3",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
repops-api    = "repops.api.main:run"
repops-worker = "repops.workers.app:run"

[tool.hatch.build.targets.wheel]
packages = ["repops"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length    = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["repops"]

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version      = "3.11"
strict              = true
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
addopts      = "-v --tb=short"
